test: share setup between tests
