+++ tests/e2e/jest.setup.js
@@ -1,0 +1,7 @@
+jest.mock("serialport", () => {
+    const actual = jest.requireActual("serialport");
+    return {
+        ...actual,
+        SerialPort: actual.SerialPortMock
+    };
+});
